refactoring on the fly
